Django/Books/Django.for.Professionals/bookstore/requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==5.0.7
django-allauth==0.63.6
django-cache-url==3.4.5
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
environs==11.0.0
marshmallow==3.21.3
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
